pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drf-auth-email"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "ArtyomYaprintsev", email = "yapryntsev.a02@mail.ru" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "djangorestframework>=3.0",
  "django-ipware>=5.0",
]

[project.urls]
Documentation = "https://github.com/ArtyomYaprintsev/drf-auth-email#readme"
Issues = "https://github.com/ArtyomYaprintsev/drf-auth-email/issues"
Source = "https://github.com/ArtyomYaprintsev/drf-auth-email"

[tool.hatch.version]
path = "src/drf_auth_email/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/drf_auth_email tests}"

[tool.coverage.run]
source_pkgs = ["drf_auth_email", "tests"]
branch = true
parallel = true
omit = [
  "src/drf_auth_email/__about__.py",
]

[tool.coverage.paths]
drf_auth_email = ["src/drf_auth_email", "*/drf-auth-email/src/drf_auth_email"]
tests = ["tests", "*/drf-auth-email/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
